misc/HtmlTestSuite/results/test-1104.xhtml

<?xml version="1.0" encoding="UTF-8"?><html xmlns="http://www.w3.org/1999/xhtml"><head><meta charset="utf-8"/>
<title>Doctype without root name should have empty-string name in the DOM even if null in the tokenizer spec.</title>
<script src="/resources/testharness.js"/>
<script src="/resources/testharnessreport.js"/>
<script>
setup({explicit_done:true});
window.onload = function() {
  test(function () {
    assert_equals(document.doctype.name, "", "Top-level");
    let iframes = document.getElementsByTagName("iframe");
    for (let i = 0; i &lt; iframes.length; ++i) {
      let iframe = iframes[i];
      assert_equals(iframe.contentDocument.doctype.name, "", iframe.title);
    }
  }, "Doctype without root name should have the empty string as the name.");
  done();
}
</script>
</head><body><iframe src="support/no-doctype-name-space.html" title="space"/>
<iframe src="support/no-doctype-name-line.html" title="line"/>
<iframe src="support/no-doctype-name-eof.html" title="eof"/>
</body></html>
